refactor : s3 업로드 로직 분리
