+++ croffle/src/main/java/com/be/croffle/music/MusicGenService.java
@@ -29,13 +29,8 @@ import java.util.List;
 @RequiredArgsConstructor
 @Transactional
 public class MusicGenService {
-
-    private final AmazonS3Client amazonS3Client;
     private final MusicJpaRepository musicJpaRepository;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
+    private final S3uploader s3uploader;
     private final String workingDir = System.getProperty("user.dir");
 
     //TODO : 환경변수화
@@ -51,7 +46,7 @@ public class MusicGenService {
         JSONObject jsonObject = createJsonFromRequest(reqDto);
         String uuid = executePythonScript(jsonObject.toString());
 
-        String musicUrl = uploadFileToS3(uuid);
+        String musicUrl = s3uploader.uploadFileToS3(uuid);
         musicJpaRepository.save(Music
                 .builder()
                 .musicUrl(musicUrl)
@@ -94,8 +89,6 @@ public class MusicGenService {
         return null;
     }
 
-
-
     private void logProcessOutput(BufferedReader reader) throws IOException {
         String output = reader.lines().reduce("", (acc, line) -> acc + "\n" + line);
         log.info("Python script output: {}", output);
@@ -106,34 +99,7 @@ public class MusicGenService {
         log.info("Python script exited with code {}", exitCode);
     }
 
-    private String uploadFileToS3(String musicName) {
-        // AWS S3 업로드 로직 구현
-        //TODO : 음악 확장자 확인
-        File uploadFile =  new File(System.getProperty("user.dir") + "/musics/" + "music" + musicName +".wav");
-
-        String fileName = "music" + "/" + uploadFile.getName(); // S3에 저장된 파일 이름
-        String uploadMusicUrl = putS3(uploadFile, fileName); // S3로 업로드
-        log.info("uploadImageUrl = " + uploadMusicUrl);
-        removeNewFile(uploadFile);
-
-        // 업로드된 음악 url 제공
-        return uploadMusicUrl;
-    }
-
-    private String putS3(File uploadFile, String fileName) {
-        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile)
-                .withCannedAcl(CannedAccessControlList.PublicRead));
-        return amazonS3Client.getUrl(bucket, fileName).toString();
-    }
-
-    private void removeNewFile(File targetFile) {
-        if (targetFile.delete()) {
-            log.info("파일이 삭제되었습니다.");
-        } else {
-            log.info("파일이 삭제되지 못했습니다.");
-        }
-    }
-
+    @Transactional(readOnly = true)
     public PlaylistResponse getPlaylist() {
 
         Pageable pageable = PageRequest.of(0, 10, Sort.by("id").descending());
